requirements.txt
fastapi
uvicorn
langgraph
psycopg[binary]
openai
httpx
python-dotenv
